requirements.txt
alembic==1.16.2
blinker==1.9.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.3.0
pandas==2.3.0
pillow==11.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
qrcode==8.2
reportlab==4.4.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
tzdata==2025.2
Werkzeug==3.1.3
WTForms==3.2.1
xlsxwriter==3.2.5
